fix update of version

+++ c/sw_packages/core/version_management/Makefile.uc.mk
@@ -1,6 +1,8 @@
 
 INCLUDE_THIS_COMPONENT := $(CONFIG_INCLUDE_VERSION_MENAGMENT)
 
+include $(MAKEFILES_ROOT_DIR)/_functions/usefull_functions.mk
+
 
 #INCLUDE_DIR = 
 
@@ -11,11 +13,12 @@ INCLUDE_THIS_COMPONENT := $(CONFIG_INCLUDE_VERSION_MENAGMENT)
 #ASMFLAGS =  
 
 # force this component to be compiled always to get updated version
-CURRENT_COMPILATION_DIR_NAME := $(notdir $(abspath .))
-CURR_OBJ_DIR :=$(call fix_path_if_in_windows,\
-                 $(OBJ_DIR)/$(CURRENT_COMPILATION_DIR_NAME))
-SHELL_OUT :=$(shell $(RM) $(CURR_OBJ_DIR) 2>&1)
-
+ifeq ($(MAKECMDGOALS),all)
+    CURRENT_COMPILATION_DIR_NAME := $(notdir $(abspath .))
+    CURR_OBJ_DIR :=$(call fix_path_if_in_windows,\
+                    $(OBJ_DIR)/$(CURRENT_COMPILATION_DIR_NAME))
+    SHELL_OUT :=$(shell $(RM) $(CURR_OBJ_DIR) 2>&1)
+endif
 
 DEFINES += VERSION_STR="$(FULL_VERSION_STR)"
